Big refactoring of the IT Modernization data (#7)
This is a large change of the IT modernization to support multiple source files and a more modular approach for combining different types of data. I expect to keep iterating on this but the goal is to make the root-level files consistent / more stable while I might modify back-end files (need to work on schema metadata, etc. next)